Split up interface into multiple files
