--- firedrake/layer_potentials.py
@@ -1,1226 +1,0 @@
-import numpy as np
-import firedrake.function
-
-from firedrake.functionspaceimpl import WithGeometry
-from firedrake.mesh import MeshGeometry
-from firedrake.pointwise_operators import AbstractExternalOperator
-from firedrake.utils import cached_property
-
-from firedrake import Function, FunctionSpace, interpolate, Interpolator, \
-    SpatialCoordinate, VectorFunctionSpace, TensorFunctionSpace, project
-
-from ufl.algorithms.apply_derivatives import VariableRuleset
-from ufl.algorithms import extract_coefficients
-from ufl.constantvalue import as_ufl
-from ufl.core.multiindex import indices
-from ufl.tensors import as_tensor
-
-from pyop2.datatypes import ScalarType
-from warnings import warn
-
-
-__all__ = ("DoubleLayerPotential", "PytentialLayerOperation",
-           "SingleLayerPotential",
-           "SingleOrDoubleLayerPotential",  # included for the docs
-           "VolumePotential",)
-
-
-class Potential(AbstractExternalOperator):
-    r"""
-    This is a function which represents a potential
-    computed on a firedrake function.
-
-    For a firedrake function :math:`u:\Omega\to\mathbb C^n`
-    with :math:`\Omega \subseteq \mathbb R^m`,
-    the potential :math:`P(u):\Omega \to\mathbb C^n` is a
-    convolution of :math:`u` against some kernel function.
-    More concretely, given a source region :math:`\Gamma \subseteq \Omega`
-    a target region :math:`\Sigma \subseteq \Omega`, and
-    a kernel function :math:`K`, we define
-
-    .. math::
-
-        P(u)(x) = \begin{cases}
-            \int_\Gamma K(x, y) u(y) \,dy & x \in \Sigma
-            0 & x \notin \Sigma
-        \end{cases}
-    """
-    _external_operator_type = 'GLOBAL'
-
-    def __init__(self, density, **kwargs):
-        """
-        :arg density: A :mod:`firedrake` :class:`firedrake.function.Function`
-            or UFL expression which represents the density :math:`u`
-        :kwarg connection: A :class:`PotentialEvaluationLibraryConnection`
-        :kwarg potential_operator: The external potential evaluation library
-                                 bound operator.
-        """
-        # super
-        AbstractExternalOperator.__init__(self, density, **kwargs)
-
-        # FIXME
-        for order in self.derivatives:
-            assert order == 1, "Assumes self.derivatives = (1,..,1)"
-
-        # Get connection & bound op and validate
-        connection = kwargs.get("connection", None)
-        if connection is None:
-            raise ValueError("Missing kwarg 'connection'")
-        if not isinstance(connection, PotentialEvaluationLibraryConnection):
-            raise TypeError("connection must be of type "
-                            "PotentialEvaluationLibraryConnection, not %s."
-                            % type(connection))
-        self.connection = connection
-
-        self.potential_operator = kwargs.get("potential_operator", None)
-        if self.potential_operator is None:
-            raise ValueError("Missing kwarg 'potential_operator'")
-
-        # Get function space and validate aginst bound op
-        function_space = self.ufl_function_space()
-        assert isinstance(function_space, WithGeometry)  # sanity check
-        if function_space is not self.potential_operator.function_space():
-            raise ValueError("function_space must be same object as "
-                             "potential_operator.function_space().")
-
-        # Make sure density is a member of our function space, if it is
-        if isinstance(density, Function):
-            if density.function_space() is not function_space:
-                raise ValueError("density.function_space() must be the "
-                                 "same as function_space")
-        # Make sure the shapes match, at least
-        elif density.shape != function_space.shape:
-            raise ValueError("Shape mismatch between function_space and "
-                             "density. %s != %s." %
-                             (density.shape, function_space.shape))
-
-    @cached_property
-    def _evaluator(self):
-        return PotentialEvaluator(self,
-                                  self.density,
-                                  self.connection,
-                                  self.potential_operator)
-
-    def _evaluate(self):
-        raise NotImplementedError
-
-    def _compute_derivatives(self, continuity_tolerance=None):
-        raise NotImplementedError
-
-    def _evaluate_action(self, *args):
-        return self._evaluator._evaluate_action()
-
-    def _evaluate_adjoint_action(self, *args):
-        return self._evaluator._evaluate_action()
-
-    def evaluate_adj_component_control(self, x, idx):
-        derivatives = tuple(dj + int(idx == j) for j, dj in enumerate(self.derivatives))
-        dN = self._ufl_expr_reconstruct_(*self.ufl_operands, derivatives=derivatives)
-        return dN._evaluate_adjoint_action((x.function,)).vector()
-
-    def evaluate_adj_component(self, x, idx):
-        print(x, type(x))
-        raise NotImplementedError
-
-
-class PotentialEvaluator:
-    """
-    Evaluates a potential
-    """
-    def __init__(self, density, connection, potential_operator):
-        """
-        :arg density: The UFL/firedrake density function
-        :arg connection: A :class:`PotentialEvaluationLibraryConnection`
-        :arg potential_operator: The external potential evaluation library
-                                 bound operator.
-        """
-        self.density = density
-        self.connection = connection
-        self.potential_operator = potential_operator
-
-    def _eval_potential_operator(self, action_coefficients, out=None):
-        """
-        Evaluate the potential on the action coefficients and return.
-        If *out* is not *None*, stores the result in *out*
-
-        :arg action_coefficients: A :class:`~firedrake.function.Function`
-                                  to evaluate the potential at
-        :arg out: If not *None*, then a :class:`~firedrake.function.Function`
-                  in which to store the evaluated potential
-        :return: *out* if it is not *None*, otherwise a new
-                 :class:`firedrake.function.Function` storing the evaluated
-                 potential
-        """
-        density = self.connection.from_firedrake(action_coefficients)
-        potential = self.potential_operator(density)
-        return self.connection.to_firedrake(potential, out=out)
-
-    def _evaluate(self):
-        """
-        Evaluate P(self.density) into self
-        """
-        return self._eval_potential_operator(self.density, out=self)
-
-    def _compute_derivatives(self):
-        """
-        Return a function
-        Derivative(P, self.derivatives, self.density)
-        """
-        # FIXME : Assumes derivatives are Jacobian
-        return self._eval_potential_operator
-
-    def _evaluate_action(self, action_coefficients):
-        """
-        Evaluate derivatives of layer potential at action coefficients.
-        i.e. Derivative(P, self.derivatives, self.density) evaluated at
-        the action coefficients and store into self
-        """
-        operator = self._compute_derivatives()
-        return operator(action_coefficients, out=self)
-
-
-class PotentialEvaluationLibraryConnection:
-    """
-    A connection to an external library for potential evaluation
-    """
-    def __init__(self, function_space, potential_source_and_target,
-                 warn_if_cg=True):
-        """
-        Initialize self to work on the function space
-
-        :arg function_space: The :mod:`firedrake` function space
-            (of type :class:`~firedrake.functionspaceimpl.WithGeometry`) on
-            which to convert to/from. Must be a 'Lagrange' or
-            'Discontinuous Lagrange' space.
-        :arg potential_source_and_target: A :class:`PotentialSourceAndTarget`.
-            mesh must match that of function_space.
-        :arg warn_if_cg: If *True*, warn if the space is "CG"
-        """
-        # validate function space
-        if not isinstance(function_space, WithGeometry):
-            raise TypeError("function_space must be of type WithGeometry, not "
-                            "%s" % type(function_space))
-
-        family = function_space.ufl_element().family()
-        acceptable_families = ['Discontinuous Lagrange', 'Lagrange']
-        if family not in acceptable_families:
-            raise ValueError("function_space.ufl_element().family() must be "
-                             "one of %s, not '%s'" %
-                             (acceptable_families, family))
-        if family == 'Lagrange' and warn_if_cg:
-            warn("Functions in continuous function space will be projected "
-                 "to/from a 'Discontinuous Lagrange' space. Make sure "
-                 "any operators evaluated are continuous. "
-                 "Pass warn_if_cg=False to suppress this warning.")
-
-        # validate potential_source_and_targets
-        if not isinstance(potential_source_and_target,
-                          PotentialSourceAndTarget):
-            raise TypeError("potential_source_and_targets must be of type "
-                            "PotentialSourceAndTarget, not '%s'."
-                            % type(potential_source_and_target))
-
-        # make sure meshes match
-        if potential_source_and_target.mesh is not function_space.mesh():
-            raise ValueError("function_space.mesh() and "
-                             "potential_source_and_target.mesh must be the same"
-                             " obejct.")
-
-        # build DG space if necessary
-        family = function_space.ufl_element().family()
-        if family == 'Discontinuous Lagrange':
-            dg_function_space = function_space
-        elif family == 'Lagrange':
-            mesh = function_space.mesh()
-            degree = function_space.ufl_element().degree()
-            shape = function_space.shape
-            if shape is None or len(shape) == 0:
-                dg_function_space = FunctionSpace(mesh, "DG", degree)
-            elif len(shape) == 1:
-                dg_function_space = VectorFunctionSpace(mesh, "DG", degree,
-                                                        dim=shape)
-            else:
-                dg_function_space = TensorFunctionSpace(mesh, "DG", degree,
-                                                        shape=shape)
-        else:
-            acceptable_families = ['Discontinuous Lagrange', 'Lagrange']
-            raise ValueError("function_space.ufl_element().family() must be "
-                             "one of %s, not '%s'" %
-                             (acceptable_families, family))
-
-        # store function space and dg function space
-        self.function_space = function_space
-        self.dg_function_space = dg_function_space
-        self.is_dg = function_space == dg_function_space
-        # store source and targets
-        self.source_and_target = potential_source_and_target
-
-    def from_firedrake(self, density):
-        """
-        Convert the density into a form acceptable by an bound operation
-        in an external library
-
-        :arg density: A :class:`~firedrake.function.Function` holding the
-                      density.
-        :returns: The converted density
-        """
-        raise NotImplementedError
-
-    def to_firedrake(self, evaluated_potential, out=None):
-        """
-        Convert the evaluated potential from an external library
-        into a firedrake function
-
-        :arg evaluated_potential: the evaluated potential
-        :arg out: If not *None*, store the converted potential into this
-                  :class:`firedrake.function.Function`.
-        :return: *out* if it is not *None*, otherwise a new
-                 :class:`firedrake.function.Function` storing the evaluated
-                 potential
-        """
-        raise NotImplementedError
-
-
-class MeshmodeConnection(PotentialEvaluationLibraryConnection):
-    """
-    Build a connection into :mod:`meshmode`
-    """
-    def __init__(self, function_space, potential_source_and_target, actx,
-                 warn_if_cg=True,
-                 meshmode_connection_kwargs=None):
-        """
-        For other args see :class:`PotentialEvaluationLibraryConnection`
-
-        :arg actx: a :class:`meshmode.array_context.PyOpenCLArrayContext`
-           used for :func:`meshmode.interop.build_connection_from_firedrake`
-           and for conversion by the
-           :class:`meshmode.interop.FiredrakeConnection`.
-        :arg meshmode_connection_kwargs: A dict passed as kwargs
-            to :func:`meshmode.interop.build_connection_from_firedrake`.
-            Must not include 'restrict_to_boundary'
-        """
-        PotentialEvaluationLibraryConnection.__init__(
-            self,
-            function_space,
-            potential_source_and_target,
-            warn_if_cg=warn_if_cg)
-
-        # FIXME : allow subdomain regions
-        tdim = potential_source_and_target.mesh.topological_dimension()
-        if potential_source_and_target.get_source_dim() != tdim:
-            if potential_source_and_target.get_source_id() != "everywhere":
-                raise NotImplementedError("subdomain sources not implemented")
-        if potential_source_and_target.get_target_dim() != tdim:
-            if potential_source_and_target.get_target_id() != "everywhere":
-                raise NotImplementedError("subdomain targets not implemented")
-
-        # validate actx
-        from meshmode.array_context import PyOpenCLArrayContext
-        if not isinstance(actx, PyOpenCLArrayContext):
-            raise TypeError("actx must be of type PyOpenCLArrayContext, not "
-                            "%s." % type(actx))
-
-        # validate meshmode_connection_kwargs
-        if meshmode_connection_kwargs is None:
-            meshmode_connection_kwargs = {}
-        if not isinstance(meshmode_connection_kwargs, dict):
-            raise TypeError("meshmode_connection_kwargs must be *None* or of "
-                            "type dict, not '%s'." %
-                            type(meshmode_connection_kwargs))
-        if 'restrict_to_boundary' in meshmode_connection_kwargs:
-            raise ValueError("'restrict_to_boundary' must not be set by "
-                             "meshmode_connection_kwargs")
-
-        # build a meshmode connection for the source
-        src_bdy = None
-        if potential_source_and_target.get_source_dim() != tdim:
-            # sanity check
-            assert potential_source_and_target.get_source_dim() == tdim - 1
-            src_bdy = potential_source_and_target.get_source_id()
-
-        from meshmode.interop.firedrake import build_connection_from_firedrake
-        src_conn = build_connection_from_firedrake(actx,
-                                                   self.dg_function_space,
-                                                   restrict_to_boundary=src_bdy,
-                                                   **meshmode_connection_kwargs)
-        # If source is a boundary, build a connection to restrict it to the
-        # boundary
-        restrict_conn = None
-        if src_bdy is not None:
-            # convert "everywhere" to meshmode BTAG_ALL
-            meshmode_src_bdy = src_bdy
-            if meshmode_src_bdy == "everywhere":
-                from meshmode.mesh import BTAG_ALL
-                meshmode_src_bdy = BTAG_ALL
-            # get group factory
-            order = self.dg_function_space.degree()
-            from meshmode.discretization.poly_element import \
-                PolynomialRecursiveNodesGroupFactory
-            default_grp_factory = \
-                PolynomialRecursiveNodesGroupFactory(order, 'lgl')
-            grp_factory = meshmode_connection_kwargs.get('grp_factory',
-                                                         default_grp_factory)
-            from meshmode.discretization.connection import make_face_restriction
-            restrict_conn = make_face_restriction(actx,
-                                                  src_conn.discr,
-                                                  grp_factory,
-                                                  meshmode_src_bdy)
-
-        # Build a meshmode connection for the target
-        tgt_bdy = None
-        if potential_source_and_target.get_target_dim() != tdim:
-            # sanity check
-            assert potential_source_and_target.get_target_dim() == tdim - 1
-            tgt_bdy = potential_source_and_target.get_target_id()
-
-        # Can we re-use the source connection?
-        src_dim = potential_source_and_target.get_source_dim()
-        tgt_dim = potential_source_and_target.get_target_dim()
-        if tgt_bdy == src_bdy and src_dim == tgt_dim:
-            tgt_conn = src_conn
-        else:
-            # If not, build a new one
-            tgt_conn = build_connection_from_firedrake(
-                actx,
-                self.dg_function_space,
-                restrict_to_boundary=tgt_bdy,
-                **meshmode_connection_kwargs)
-
-        # store computing context
-        self.actx = actx
-        # store connections
-        self.source_to_meshmode_connection = src_conn
-        self.restrict_source_to_boundary = restrict_conn
-        self.target_to_meshmode_connection = tgt_conn
-
-    def from_firedrake(self, density):
-        # make sure we are in the dg function space
-        if not self.is_dg:
-            density = project(density, self.dg_function_space)
-        # Convert to meshmode.
-        density = \
-            self.source_to_meshmode_connection.from_firedrake(density,
-                                                              actx=self.actx)
-        # restrict to boundary if necessary
-        if self.restrict_source_to_boundary is not None:
-            density = self.restrict_source_to_boundary(density)
-        return density
-
-    def to_firedrake(self, evaluated_potential, out=None):
-        # if we are dg, it's simple
-        if self.is_dg:
-            return self.target_to_meshmode_connection.from_meshmode(
-                evaluated_potential,
-                out=out)
-        else:
-            # Otherwise, we have to project back to our function space
-            pot = \
-                self.target_to_meshmode_connection.from_meshmode(evaluated_potential)
-            pot = project(evaluated_potential, self.function_space)
-            if out is not None:
-                out.dat.data[:] = pot.dat.data[:]
-                pot = out
-            return pot
-
-    def get_source_discretization(self):
-        """
-        Get the :class:`meshmode.discretization.Discretization`
-        of the source
-        """
-        if self.restrict_source_to_boundary is None:
-            return self.source_to_meshmode_connection.discr
-        return self.restrict_source_to_boundary.to_discr
-
-    def get_target_discretization(self):
-        """
-        Get the :class:`meshmode.discretization.Discretization`
-        of the target
-        """
-        return self.target_to_meshmode_connection.discr
-
-
-class VolumentialConnection(PotentialEvaluationLibraryConnection):
-    """
-    Build a connection into :mod:`volumential`
-    """
-    def __init__(self,
-                 potential_source_and_target,
-                 function_space,
-                 cl_ctx,
-                 queue,
-                 box_fmm_geometry_kwargs,
-                 reorder_potentials,
-                 meshmode_connection_kwargs,
-                 warn_if_cg=True):
-        """
-        For other args see :class:`PotentialEvaluationLibraryConnection`.
-        The *potential_source_and_target* must have a 3D
-        source, 3D target, and 3D mesh of co-dimension 0.
-
-        :arg cl_ctx: A :class:`pyopencl.Context`
-        :arg queue: A :class:`pyopencl.CommandQueue`
-        :arg box_fmm_geometry_kwargs: kwargs passed to
-            :class:`volumential.geometry.BoxFMMGeometryFactory`.
-            Must include the key *'nlevels'*
-        :arg reorder_potentials: *True* iff the volumential FMM
-            is reordering potentials. *False* otherwise.
-        :arg meshmode_connection_kwargs: A dict passes as kwargs
-            to :func:`meshmode.interop.build_connection_from_firedrake`.
-            :mod:`meshmode` is used an intermediate level between
-            :mod:`firedrake` and :mod:`volumential`.
-        """
-        # super
-        PotentialEvaluationLibraryConnection.__init__(
-            self,
-            function_space,
-            potential_source_and_target,
-            warn_if_cg=warn_if_cg)
-
-        # validate that sources, target, and mesh are of right dimension
-        if potential_source_and_target.get_source_dimension() != 3:
-            raise ValueError("potential source must have dimension 3")
-        if potential_source_and_target.get_target_dimension() != 3:
-            raise ValueError("potential target must have dimension 3")
-        if potential_source_and_target.mesh().geometric_dimension() != 3:
-            raise ValueError("mesh must have geometric dimension 3")
-        if potential_source_and_target.mesh().topological_dimension() != 3:
-            raise ValueError("mesh must have topological dimension 3")
-
-        # FIXME : allow subdomain regions
-        if potential_source_and_target.get_source_id() != "everywhere":
-            raise NotImplementedError("subdomain sources not implemented")
-        if potential_source_and_target.get_target_id() != "everywhere":
-            raise NotImplementedError("subdomain targets not implemented")
-
-        # validate cl_ctx argument
-        from pyopencl import Context
-        if not isinstance(cl_ctx, Context):
-            raise TypeError("cl_ctx must be of type Context, not "
-                            "%s." % type(cl_ctx))
-
-        # validate queue argument
-        from pyopencl import CommandQueue
-        if not isinstance(queue, CommandQueue):
-            raise TypeError("queue must be of type CommandQueue, not "
-                            "%s." % type(queue))
-
-        # validate reorder_potentials
-        if not isinstance(reorder_potentials, bool):
-            raise TypeError("reorder_potentials must be of type bool, not "
-                            "%s." % type(reorder_potentials))
-
-        # validate box_fmm_geometry_kwargs
-        if not isinstance(box_fmm_geometry_kwargs, dict):
-            raise TypeError("box_fmm_geometry_kwargs must be of "
-                            "type dict, not '%s'." %
-                            type(box_fmm_geometry_kwargs))
-        if 'nlevels' not in box_fmm_geometry_kwargs:
-            raise ValueError("box_fmm_geometry_kwargs missing required keyword"
-                             " 'nlevels'.")
-
-        # validate meshmode_connection_kwargs
-        if meshmode_connection_kwargs is None:
-            meshmode_connection_kwargs = {}
-        if not isinstance(meshmode_connection_kwargs, dict):
-            raise TypeError("meshmode_connection_kwargs must be *None* or of "
-                            "type dict, not '%s'." %
-                            type(meshmode_connection_kwargs))
-
-        # Build intermediate connection into meshmode
-        from meshmode.interop.firedrake import build_connection_from_firedrake
-        from meshmode.array_context import PyOpenCLArrayContext
-        actx = PyOpenCLArrayContext(queue)
-        meshmode_connection = MeshmodeConnection(
-                function_space,
-                potential_source_and_target,
-                actx,
-                warn_if_cg=warn_if_cg,
-                meshmode_connection_kwargs=meshmode_connection_kwargs)
-
-        # Build connection from meshmode into volumential
-        # (following https://gitlab.tiker.net/xywei/volumential/-/blob/fe2c3e7af355d5c527060e783237c124c95397b5/test/test_interpolation.py#L72 ) # noqa : E501
-        from volumential.geometry import (
-            BoundingBoxFactory, BoxFMMGeometryFactory)
-        from volumential.interpolation import ElementsToSourcesLookupBuilder
-        dim = function_space.mesh().geometric_dimension()
-        bboxfmm_fac = BoxFMMGeometryFactory(cl_ctx, dim=dim,
-                                            **box_fmm_geometry_kwargs)
-        boxgeo = bboxfmm_fac(queue)
-        elt_to_src_lookup_fac = ElementsToSourcesLookupBuilder(
-            cl_ctx, tree=boxgeo.tree, discr=meshmode_connection.discr)
-        elt_to_src_lookup, evt = elt_to_src_lookup_fac(queue)
-
-        # Build connection from volumential into meshmode
-        from volumential.interpolation import LeavesToNodesLookupBuilder
-        leaves_to_node_lookup_fac = LeavesToNodesLookupBuilder(
-            cl_ctx, trav=boxgeo.trav, discr=meshmode_connection.discr)
-        leaves_to_node_lookup, evt = leaves_to_node_lookup_fac(queue)
-
-        # Figure out whether or not conversion from volumential -> meshmode
-        # should user order or tree order
-        order = "user" if reorder_potentials else "tree"
-
-        # Store maps for conversion to/from volumential
-        self.meshmode_connection = meshmode_connection
-        self.to_volumential_lookup = elt_to_src_lookup
-        self.from_volumential_lookup = leaves_to_node_lookup
-        self.from_volumential_order = order
-        # store necessary computing contextx
-        self.queue = queue
-        self.actx = actx
-
-    def from_firedrake(self, density):
-        # pass operand into a meshmode DOFArray
-        from meshmode.dof_array import flatten
-        meshmode_density = self.meshmode_connection.from_firedrake(density)
-        meshmode_density = flatten(meshmode_density)
-
-        # pass flattened operand into volumential interpolator
-        from volumential.interpolation import interpolate_from_meshmode
-        volumential_density = \
-            interpolate_from_meshmode(self.queue,
-                                      meshmode_density,
-                                      self.to_volumential_lookup,
-                                      order="user")  # user order is more intuitive
-        return volumential_density
-
-    def to_firedrake(self, evaluated_potential, out=None):
-        # pass volumential back to meshmode DOFArray
-        from volumential.interpolation import interpolate_to_meshmode
-        from meshmode.dof_array import unflatten
-        meshmode_pot_vals = \
-            interpolate_to_meshmode(self.queue,
-                                    evaluated_potential,
-                                    self.from_volumential_lookup,
-                                    order=self.from_volumential_order)
-        meshmode_pot_vals = unflatten(self.actx,
-                                      self.meshmode_connection.discr,
-                                      meshmode_pot_vals)
-        # get meshmode data back as firedrake function
-        return self.meshmode_connection.to_firedrake(meshmode_pot_vals, out=self)
-
-
-"""
-functions
-    SingleLayerPotential
-    DoubleLayerPotential
-    VolumentialOperation
-    VolumePotential
-"""
-
-
-class PotentialSourceAndTarget:
-    """
-    Holds the source and target for a layer or volume potential
-    """
-    def __init__(self, mesh,
-                 source_region_dim=None,
-                 source_region_id=None,
-                 target_region_dim=None,
-                 target_region_id=None):
-        """
-        Source and target of a layer or volume potential.
-        The mesh must have co-dimension 0 or 1
-
-        By region_dim, we mean the topological dimension of the
-        region.
-        Regions must have co-dimensions 0 or 1.
-        They cannot have a higher dimension than the mesh.
-        *None* indicates the topological dimension of the
-        mesh.
-
-        Region ids must be either a valid mesh subdomain id (an
-        integer or tuple) or *None*. *None* indicates either
-        the entire mesh, or the entire exterior boundary
-        (as determined by the value of region).
-        The string "everywhere" is also equivalent to a
-        value of *None*
-
-        NOTE: For implementation reasons, if the target is
-              a boundary, instead of just evaluating the
-              target at DOFs on that boundary, it is instead
-              evaluated at any DOF of a cell which has at least
-              one vertex on the boundary.
-        """
-        # validate mesh
-        if not isinstance(mesh, MeshGeometry):
-            raise TypeError("mesh must be of type MeshGeometry, not '%s'." %
-                            type(mesh))
-
-        # mesh must have co-dimension 1 or 0.
-        valid_codims = [0, 1]
-        codim = mesh.geometric_dimension() - mesh.topological_dimension()
-        if codim not in valid_codims:
-            raise ValueError("mesh has invalid co-dimension of %s. "
-                             "co-dimension must be one of %s." %
-                             (codim, valid_codims))
-
-        # validate dims
-        for dim in [source_region_dim, target_region_dim]:
-            if dim is None:
-                continue
-            if not isinstance(dim, int):
-                raise TypeError("source and target dimensions must be *None*"
-                                " or an integer, not of type '%s'." % type(dim))
-            valid_dims = set([mesh.topological_dimension(),
-                              mesh.geometric_dimension()-1])
-            if dim < mesh.geometric_dimension() - 1:
-                raise ValueError("source and target dimensions must be "
-                                 "one of %s or *None*, not '%s'." %
-                                 (valid_dims, dim))
-        # Get dims if *None*
-        if source_region_dim is None:
-            source_region_dim = mesh.topological_dimension()
-        if target_region_dim is None:
-            target_region_dim = mesh.topological_dimension()
-
-        # Validate region ids type and validity
-        for id_, dim in zip([source_region_id, target_region_id],
-                            [source_region_dim, target_region_dim]):
-            if id_ is None or id_ == "everywhere":
-                continue
-            # standardize id_ to a tuple
-            if isinstance(id_, int):
-                id_ = tuple(id_)
-            # check type
-            if not isinstance(id_, tuple):
-                raise TypeError("source and target region ids must be "
-                                "*None*, an int, or a tuple, not of type")
-            # boundary case:
-            if dim == mesh.topological_dimension() - 1:
-                if not set(id_) <= set(mesh.exterior_facets.unique_markers):
-                    raise ValueError(("boundary region ids %s are not a "
-                                     + "subset of mesh.exterior_facets."
-                                     + "unique_markers = %s.") % (id_,
-                                     mesh.exterior_facets.unique_markers))
-            else:
-                # sanity check
-                assert dim == mesh.topological_dimension()
-                # this caches the cell subset on the mesh, which we'll
-                # probably be doing anyway if we're targeting the region
-                #
-                # It also throws a ValueError if the id_ is invalid
-                mesh.cell_subset(id_)
-
-        # handle None case
-        if source_region_id is None:
-            source_region_id = "everywhere"
-        if target_region_id is None:
-            target_region_id = "everywhere"
-
-        # store mesh, region dims, and region ids
-        self.mesh = mesh
-        self._source_region_dim = source_region_dim
-        self._target_region_dim = source_region_dim
-        self._source_region_id = source_region_id
-        self._target_region_id = source_region_id
-
-    def get_source_dimension(self):
-        """
-        Get the topological dimension of the source region
-        """
-        return self._source_region_dim
-
-    def get_target_dimension(self):
-        """
-        Get the topological dimension of the target region
-        """
-        return self._target_region_dim
-
-    def get_source_id(self):
-        """
-        Get the subdomain id of the source, or the string
-        "everywhere" to represent the entire exterior boundary/mesh
-        """
-        return self._source_region_id
-
-    def get_target_id(self):
-        """
-        Get the subdomain id of the target, or the string
-        "everywhere" to represent the entire exterior boundary/mesh
-        """
-        return self._target_region_id
-
-
-def PytentialOperation(actx,
-                       density,
-                       unbound_op,
-                       density_name,
-                       potential_src_and_tgt,
-                       **kwargs):
-    """
-    :arg actx: A :class:`meshmode.dof_array.PyOpenCLArrayContext`
-    :arg density: the :mod:`firedrake` density function
-    :arg unbound_op: A :mod:`pytential` operation which has not been
-        bound (e.g. a :mod:`pymbolic` expression)
-    :arg density_name: A string, the name of the density function
-        in the unbound_op
-    :arg potential_src_and_tgt: A :class:`PotentialSourceAndTarget`
-
-    :kwarg warn_if_cg: Pass *False* to suppress the warning if
-        a "CG" space is used
-    :kwarg meshmode_connection_kwargs: *None*, or
-        a dict with arguments to pass to
-        :func:`meshmode.interop.firedrake.build_connection_from_firedrake`
-    :kwarg qbx_kwargs: *None*, or a dict with arguments to pass to
-        :class:`pytential.qbx.QBXLayerPotentialSource`.
-    :kwarg op_kwargs: kwargs passed to the invocation of the
-        bound pytential operator (e.g. {'k'=0.5} for
-        a :class:`sumpy.kernel.HelmholtzKernel`).
-        Must not include *density_name*
-    """
-    # make sure density name is a string
-    if not isinstance(density_name, str):
-        raise TypeError("density_name must be of type str, not '%s'." %
-                        density_name)
-
-    # get kwargs to build connection
-    warn_if_cg = kwargs.get('warn_if_cg', True)
-    meshmode_connection_kwargs = kwargs.get('meshmode_connection_kwargs', None)
-
-    # Build meshmode connection
-    meshmode_connection = MeshmodeConnection(
-            density.function_space(),
-            potential_src_and_tgt,
-            actx,
-            warn_if_cg=warn_if_cg,
-            meshmode_connection_kwargs=meshmode_connection_kwargs)
-
-    # Build QBX
-    src_discr = meshmode_connection.get_source_discretization()
-    qbx_kwargs = kwargs.get('qbx_kwargs', None)
-    from pytential.qbx import QBXLayerPotentialSource
-    qbx = QBXLayerPotentialSource(src_discr, **qbx_kwargs)
-
-    # Bind pytential operator
-    tgt_discr = meshmode_connection.get_target_discretization()
-    from pytential import bind
-    pyt_op = bind((qbx, tgt_discr), unbound_op)
-
-    # Get operator kwargs
-    op_kwargs = kwargs.get('op_kwargs', {})
-    if density_name in op_kwargs:
-        raise ValueError(f"density_name '{density_name}' should not be included"
-                         " in op_kwargs.")
-
-    # build bound operator that just takes density as argument
-    def bound_op_with_kwargs(density_arg):
-        op_kwargs[density_name] = density_arg
-        return pyt_op(**op_kwargs)
-
-    # Now build and return Potential object
-    return Potential(density,
-                     connection=meshmode_connection,
-                     potential_operator=bound_op_with_kwargs)
-
-
-def SingleLayerPotential(density, kernel, potential_src_and_tgt, **kwargs):
-    pass
-
-
-def DoubleLayerPotential(density, kernel, potential_src_and_tgt, **kwargs):
-    pass
-
-
-def VolumentialOperation(density, unbound_op, potential_src_and_tgt,
-                         function_space=None):
-    pass
-
-
-class VolumePotential(AbstractExternalOperator):
-    r"""
-    Evaluates to
-
-    .. math::
-
-         f(x) = \int_\Om K(x-y) op(y) \,dx
-
-    where op is a function-type living in *function_space*, \Om is the
-    3-D mesh on which *function_space* lives, and K
-    a :mod:`sumpy` kernel.
-
-    :arg operator_data: A map with required keys
-
-        * 'kernel': a :class:`sumpy.kernel.Kernel` (*K* in the math above)
-        * 'kernel_type': A string describing the kernel, e.g. ``"Laplace"``
-        * 'cl_ctx': a :mod:`pyopencl` computing context
-        * 'queue': a :mod:`pyopencl` command queue compatible with the
-                   computing context
-        * 'nlevels': The desired number of levels to be used by
-                      :mod:`volumential`
-        * 'm_order': multipole order used by the expansion
-        * 'dataset_filename': the filename to pass to
-            :class:`volumential.table_manager.NearFieldInteractionTableManager`
-
-        And optional keys
-
-        * 'grp_factory': (optional) An interpolatory group factory
-            inheriting from :class:`meshmode.discretization.ElementGroupFactory`
-            to be used in the intermediate :mod:`meshmode` representation
-        * 'q_order': (optional) The desired :mod:`volumential` quadrature
-                     order, defaults to *function_space*'s degree
-        * 'force_direct_evaluation': (optional) As in
-                     :func:`volumential.volume_fmm.drive_volume_fmm`.
-                     Defaults to *False*
-        * 'fmm_kwargs': (optional) A dictionary of kwargs
-                 to pass to :func:`volumential.volume_fmm.drive_fmm`.
-        * 'root_extent': (optional) the root extent to pass to
-            :class:`volumential.table_manager.NearFieldInteractionTableManager`
-            (defaults to 1)
-        * 'table_compute_method': (optional) used by
-            :class:`volumential.table_manager.NearFieldInteractionTableManager`
-            (defaults to "DrosteSum")
-        * 'table_kwargs': (optional) passed to
-            :class:`volumential.table_manager.NearFieldInteractionTableManager`
-            's *get_table* method
-
-    """
-
-    _external_operator_type = 'GLOBAL'
-
-    def __init__(self, *operands, **kwargs):
-        self.operands = operands
-        self.kwargs = kwargs
-        AbstractExternalOperator.__init__(self, operands[0], **kwargs)
-
-
-        # Build a broken version of input function space
-        fs = kwargs["function_space"]
-        fsdg = FunctionSpace(fs.mesh(), "DG", fs.ufl_element().degree())
-
-        self.evaluator_kwargs = kwargs.copy()
-        self.evaluator_kwargs["function_space"] = fsdg
-
-        # Validate input
-        # print(len(self.operands), self.derivatives)
-        # assert self.derivatives == (0,), \
-        #     "Derivatives of volume potential not currently supported: " + str(self.derivatives)
-        # from firedrake import Function
-        #
-        # if not isinstance(operand, Function):
-        #     raise TypeError(":arg:`operand` must be of type firedrake.Function"
-        #                     ", not %s." % type(operand))
-        # if operand.function_space().shape != tuple():
-        #     raise ValueError(":arg:`operand` must be a function with shape (),"
-        #                      " not %s." % operand.function_space().shape)
-        operator_data = kwargs["operator_data"]
-        assert isinstance(operator_data, dict)
-        required_keys = ('kernel', 'kernel_type', 'cl_ctx', 'queue', 'nlevels',
-                         'm_order', 'dataset_filename')
-        optional_keys = ('grp_factory', 'q_order', 'force_direct_evaluation',
-                         'fmm_kwargs', 'root_extent',
-                         'table_compute_method', 'table_kwargs',
-                         )
-        permissible_keys = required_keys + optional_keys
-        if not all(key in operator_data for key in required_keys):
-            raise ValueError("operator_data is missing one of the required "
-                             "keys: %s" % required_keys)
-        if not all(key in permissible_keys for key in operator_data):
-            raise ValueError("operator_data contains an unexpected key. All "
-                             "keys must be one of %s." % (permissible_keys,))
-
-    @cached_property
-    def _evaluator(self):
-        return VolumePotentialEvaluator(self, *self.operands, **self.evaluator_kwargs)
-
-    def _evaluate(self, continuity_tolerance=None):
-        1/0
-
-    def _compute_derivatives(self, continuity_tolerance=None):
-        1/0
-
-    def _evaluate_action(self, *args, continuity_tolerance=None):
-        return self._evaluator._evaluate_action(continuity_tolerance=continuity_tolerance)
-
-    def _evaluate_adjoint_action(self, *args, continuity_tolerance=None):
-        return self._evaluator._evaluate_action(continuity_tolerance=continuity_tolerance)
-
-    def evaluate_adj_component_control(self, x, idx):
-        derivatives = tuple(dj + int(idx == j) for j, dj in enumerate(self.derivatives))
-        dN = self._ufl_expr_reconstruct_(*self.ufl_operands, derivatives=derivatives)
-        return dN._evaluate_adjoint_action((x.function,)).vector()
-
-    def evaluate_adj_component(self, x, idx):
-        print(x, type(x))
-        1/0
-
-
-class VolumePotentialEvaluator:
-    def __init__(self, vp, *operands, **kwargs):
-        self.vp = vp
-        operand = operands[0]
-
-        cg_function_space = vp.function_space()
-        function_space = kwargs["function_space"]
-        self.dg_function_space = function_space
-
-        # Create an interpolator from the original operand so that
-        # we can re-interpolate it into the space at each evaluation
-        # in case it has changed (e.g. cost functional in PDE constrained optimization.
-
-        operator_data = kwargs["operator_data"]
-        kernel = operator_data['kernel']
-        kernel_type = operator_data['kernel_type']
-        cl_ctx = operator_data['cl_ctx']
-        queue = operator_data['queue']
-        nlevels = operator_data['nlevels']
-        m_order = operator_data['m_order']
-        dataset_filename = operator_data['dataset_filename']
-        degree = function_space.ufl_element().degree()
-        grp_factory = operator_data.get('grp_factory', None)
-        q_order = operator_data.get('q_order', degree)
-        force_direct_evaluation = operator_data.get('force_direct_evaluation',
-                                                    False)
-        fmm_kwargs = operator_data.get('fmm_kwargs', {})
-        root_extent = operator_data.get("root_extent", 1)
-        table_compute_method = operator_data.get('table_compute_method',
-                                                 'DrosteSum')
-        table_kwargs = operator_data.get('table_kwargs', {})
-
-        from sumpy.kernel import Kernel
-        if not isinstance(kernel, Kernel):
-            raise TypeError("operator_data['kernel'] must be of type "
-                            "sumpy.kernel.Kernel, not %s." % type(kernel))
-        if not isinstance(nlevels, int):
-            raise TypeError("operator_data['nlevels'] must be of type int, "
-                            "not %s." % type(nlevels))
-        if not isinstance(m_order, int):
-            raise TypeError("operator_data['m_order'] must be of type int, "
-                            "not %s." % type(m_order))
-        if not isinstance(q_order, int):
-            raise TypeError("operator_data['q_order'] must be of type int, "
-                            "not %s." % type(q_order))
-
-        # Build connection into meshmode
-        from meshmode.interop.firedrake import build_connection_from_firedrake
-        from meshmode.array_context import PyOpenCLArrayContext
-        actx = PyOpenCLArrayContext(queue)
-        meshmode_connection = build_connection_from_firedrake(
-            actx, function_space, grp_factory=grp_factory)
-
-        # Build connection from meshmode into volumential
-        # (following https://gitlab.tiker.net/xywei/volumential/-/blob/fe2c3e7af355d5c527060e783237c124c95397b5/test/test_interpolation.py#L72 ) # noqa : E501
-        from volumential.geometry import (
-            BoundingBoxFactory, BoxFMMGeometryFactory)
-        from volumential.interpolation import ElementsToSourcesLookupBuilder
-        dim = function_space.mesh().geometric_dimension()
-        bbox_fac = BoundingBoxFactory(dim=dim)
-        bboxfmm_fac = BoxFMMGeometryFactory(
-            cl_ctx, dim=dim, order=q_order,
-            nlevels=nlevels, bbox_getter=bbox_fac,
-            expand_to_hold_mesh=meshmode_connection.discr.mesh,
-            mesh_padding_factor=0.0)
-        boxgeo = bboxfmm_fac(queue)
-        elt_to_src_lookup_fac = ElementsToSourcesLookupBuilder(
-            cl_ctx, tree=boxgeo.tree, discr=meshmode_connection.discr)
-        elt_to_src_lookup, evt = elt_to_src_lookup_fac(queue)
-
-        # Build connection from volumential into meshmode
-        from volumential.interpolation import LeavesToNodesLookupBuilder
-        leaves_to_node_lookup_fac = LeavesToNodesLookupBuilder(
-            cl_ctx, trav=boxgeo.trav, discr=meshmode_connection.discr)
-        leaves_to_node_lookup, evt = leaves_to_node_lookup_fac(queue)
-
-        # Create near-field table in volumential
-        from volumential.table_manager import NearFieldInteractionTableManager
-        table_manager = NearFieldInteractionTableManager(
-            dataset_filename, root_extent=root_extent, queue=queue)
-
-        nftable, _ = table_manager.get_table(
-            dim,
-            kernel_type,
-            q_order,
-            compute_method=table_compute_method,
-            queue=queue,
-            **table_kwargs)
-
-        # Create kernel wrangler in volumential
-        from sumpy.expansion import DefaultExpansionFactory
-        out_kernels = [kernel]
-        expn_factory = DefaultExpansionFactory()
-        local_expn_class = expn_factory.get_local_expansion_class(kernel)
-        mpole_expn_class = expn_factory.get_multipole_expansion_class(kernel)
-
-        from functools import partial
-        from volumential.expansion_wrangler_fpnd import (
-            FPNDExpansionWrangler, FPNDExpansionWranglerCodeContainer)
-        wrangler_cc = FPNDExpansionWranglerCodeContainer(
-            cl_ctx,
-            partial(mpole_expn_class, kernel),
-            partial(local_expn_class, kernel),
-            out_kernels)
-        wrangler = FPNDExpansionWrangler(
-            code_container=wrangler_cc,
-            queue=queue,
-            tree=boxgeo.tree,
-            near_field_table=nftable,
-            dtype=ScalarType,
-            fmm_level_to_order=lambda kernel, kernel_args, tree, lev: m_order,
-            quad_order=q_order)
-
-        # Store attributes that we may need later
-        # self.ufl_operands = operands
-        self.actx = actx
-        self.queue = queue
-        self.sumpy_kernel = kernel
-        self.meshmode_connection = meshmode_connection
-        self.volumential_boxgeo = boxgeo
-        self.to_volumential_lookup = elt_to_src_lookup
-        self.from_volumential_lookup = leaves_to_node_lookup
-        self.force_direct_evaluation = force_direct_evaluation
-        self.fmm_kwargs = fmm_kwargs
-        self.expansion_wrangler = wrangler
-        # so we don't have to keep making a fd function during conversion
-        # from meshmode. AbstractExternalOperator s aren't functions,
-        # so can't use *self*
-        self.fd_pot = Function(self.dg_function_space)
-        self.fd_pot_cg = Function(vp.function_space())
-
-    def _evaluate(self, continuity_tolerance=None):
-        """
-        :arg continuity_tolerance: If *None* then ignored. Otherwise, a floating
-            point value. If the function space associated to this operator
-            is a continuous one and *continuity_tolerance* is a float, it is
-            verified that,
-            during the conversion of the evaluated potential
-            from meshmode's discontinuous function space representation into
-            firedrake's continuous one, the potential's value at
-            a firedrake node is within *continuity_tolerance* of its
-            value at any duplicated firedrake node
-        """
-        operand, = self.vp.ufl_operands
-        operand_discrete = interpolate(operand, self.dg_function_space)
-
-        # pass operand into a meshmode DOFArray
-        from meshmode.dof_array import flatten
-        meshmode_src_vals = \
-            self.meshmode_connection.from_firedrake(operand_discrete, actx=self.actx)
-        meshmode_src_vals = flatten(meshmode_src_vals)
-
-        # if 1:
-        #     alpha = 40
-        #     dim = 2
-
-        # if 1:
-        #     mm_discr = self.meshmode_connection.discr
-        #     from meshmode.dof_array import thaw, flat_norm
-        #     x = flatten(thaw(self.actx, mm_discr.nodes()[0]))
-        #     y = flatten(thaw(self.actx, mm_discr.nodes()[1]))
-        #     import pyopencl.clmath as clmath
-        #     import pyopencl.array as cla
-        #     norm2 = (x-0.5)**2 + (y-0.5)**2
-        #     ref_src = -(4 * alpha ** 2 * norm2 - 2 * dim * alpha) * clmath.exp(
-        #             -alpha * norm2)
-        #     denom = np.max(np.abs(ref_src.get()))
-        #     err = np.abs((ref_src - meshmode_src_vals).get())
-        #     print("Fdrake -> meshmode error:", np.max(err)/denom)
-
-        # pass flattened operand into volumential interpolator
-        from volumential.interpolation import interpolate_from_meshmode
-        volumential_src_vals = \
-            interpolate_from_meshmode(self.queue,
-                                      meshmode_src_vals,
-                                      self.to_volumential_lookup,
-                                      order="user")  # user order is more intuitive
-
-        # if 1:
-        #     # check source density against known value
-        #     # x = self.volumential_boxgeo.tree.sources[0].with_queue(self.queue)
-        #     # y = self.volumential_boxgeo.tree.sources[1].with_queue(self.queue)
-        #     x, y = self.volumential_boxgeo.nodes
-        #     import pyopencl.clmath as clmath
-        #     import pyopencl.array as cla
-        #     norm2 = (x-0.5)**2 + (y-0.5)**2
-        #     ref_src = -(4 * alpha ** 2 * norm2 - 2 * dim * alpha) * clmath.exp(
-        #             -alpha * norm2)
-        #     denom = np.max(np.abs(ref_src.get()))
-        #     err = np.abs((ref_src-volumential_src_vals).get())
-        #     print("Fdrake -> volumential error:", np.max(err)/denom)
-
-        # evaluate volume fmm
-        from volumential.volume_fmm import drive_volume_fmm
-        pot, = drive_volume_fmm(
-            self.volumential_boxgeo.traversal,
-            self.expansion_wrangler,
-            volumential_src_vals * self.volumential_boxgeo.weights,
-            volumential_src_vals,
-            direct_evaluation=self.force_direct_evaluation,
-            reorder_sources=True,
-            **self.fmm_kwargs)
-
-        # if 1:
-        #     # check potential against known value
-        #     ref_pot = clmath.exp(-alpha * norm2)
-        #     denom = np.max(np.abs(ref_pot.get()))
-        #     err = np.abs((ref_pot-pot).get())
-        #     print("volumential potential error:", np.max(err)/denom)
-
-        #     # evaluate volume fmm with ref_src
-        #     pot_w_ref_src, = drive_volume_fmm(
-        #         self.volumential_boxgeo.traversal,
-        #         self.expansion_wrangler,
-        #         ref_src * self.volumential_boxgeo.weights,
-        #         ref_src,
-        #         direct_evaluation=self.force_direct_evaluation,
-        #         reorder_sources=True,
-        #         **self.fmm_kwargs)
-
-        #     err = np.abs((ref_pot-pot_w_ref_src).get())
-        #     print("volumential potential w/t ref src error:", np.max(err)/denom)
-
-        # pass volumential back to meshmode DOFArray
-        from volumential.interpolation import interpolate_to_meshmode
-        from meshmode.dof_array import unflatten
-        user_order = self.fmm_kwargs.get("reorder_potentials", True)
-        if user_order:
-            order = "user"
-        else:
-            order = "tree"
-        meshmode_pot_vals = interpolate_to_meshmode(self.queue,
-                                                    pot,
-                                                    self.from_volumential_lookup,
-                                                    order=order)
-        meshmode_pot_vals = unflatten(self.actx,
-                                      self.meshmode_connection.discr,
-                                      meshmode_pot_vals)
-        # get meshmode data back as firedrake function
-        self.meshmode_connection.from_meshmode(
-            meshmode_pot_vals,
-            out=self.fd_pot)
-
-        self.fd_pot_cg = project(self.fd_pot, self.vp.function_space())
-
-        self.vp.dat.data[:] = self.fd_pot_cg.dat.data[:]
-
-        return self.fd_pot_cg
-
-    def _compute_derivatives(self, continuity_tolerance=None):
-        # TODO : Support derivatives
-        return self._evaluate(continuity_tolerance=continuity_tolerance)
-
-    def _evaluate_action(self, *args, continuity_tolerance=None):
-        # From tests/pointwiseoperator/test_point_expr.py
-        # https://github.com/firedrakeproject/firedrake/blob/c0d9b592f587fa8c7437f690da7a6595f6804c1b/tests/pointwiseoperator/test_point_expr.py  # noqa
-        if len(args) == 0:
-            # Evaluate the operator
-            return self._evaluate(continuity_tolerance=continuity_tolerance)
-
-        # Evaluate the Jacobian/Hessian action
-        operands = self.ufl_operands
-        operator = self._compute_derivatives(continuity_tolerance=continuity_tolerance)
-        expr = as_ufl(operator(*operands))
-        if expr.ufl_shape == () and expr != 0:
-            var = VariableRuleset(self.ufl_operands[0])
-            expr = expr*var._Id
-        elif expr == 0:
-            return self.assign(expr)
-
-        for arg in args:
-            mi = indices(len(expr.ufl_shape))
-            aa = mi
-            bb = mi[-len(arg.ufl_shape):]
-            expr = arg[bb] * expr[aa]
-            mi_tensor = tuple(e for e in mi if not (e in aa and e in bb))
-            if len(expr.ufl_free_indices):
-                expr = as_tensor(expr, mi_tensor)
-        return self.interpolate(expr)
